cnf/ran: address PR 1322 review (semver gates)
- Reject two-segment maximum strings that include a prerelease or build suffix;
  propagate NewVersion errors when constructing the implicit upper bound.
- Move IsVersionStringInRange helpers into version.go and remove semver_range.go.
- Treat explicit multi-segment maximum as the exclusive upper bound (half-open
  interval); keep plain X.Y shorthand for the next-minor exclusive bound; update
  tests plus TALM backup and PTP event-consumer call sites.
