+++ tests/cnf/ran/internal/version/version.go
@@ -3,8 +3,10 @@ package version
 import (
 	"errors"
 	"fmt"
+	"strconv"
 	"strings"
 
+	"github.com/Masterminds/semver/v3"
 	configv1 "github.com/openshift/api/config/v1"
 	"github.com/rh-ecosystem-edge/eco-goinfra/pkg/argocd"
 	"github.com/rh-ecosystem-edge/eco-goinfra/pkg/clients"
@@ -116,4 +118,150 @@ func GetZTPSiteGenerateImage(client *clients.Settings) (string, error) {
 	}
 
 	return "", errors.New("unable to identify ZTP site generate image")
+}
+
+// IsVersionStringInRange reports whether version satisfies minimum <= version < maximumUpper using SemVer 2.0
+// (github.com/Masterminds/semver/v3).
+//
+// minimum: empty means no lower bound. For a lower bound that must include pre-releases of X.Y.0 (e.g. OCP
+// 4.20.0-20251212...), pass X.Y.0-0 — the lowest pre-release of X.Y.0. Plain X.Y.0 excludes pre-releases of X.Y.0.
+//
+// maximum: empty means no upper bound. Otherwise maximum is exclusive: the interval is minimum <= v < maximum
+// (half-open). Prefer an explicit exclusive bound such as "4.18.0-0" (all 4.16.z and 4.17.z, not 4.18.0+). As a
+// shorthand only, a plain two-segment "X.Y" (no prerelease/build suffix) still means "below the next minor", i.e.
+// exclusive upper X.(Y+1).0-0 — same as passing that explicit version.
+//
+// If version is not a valid semver string and maximum is empty, the function returns (true, nil) for compatibility
+// with legacy call sites (e.g. empty or non-semver operator tags with no upper bound). Callers that require a parsed
+// version should validate the string separately or pass a non-empty maximum so the result is (false, nil).
+func IsVersionStringInRange(version, minimum, maximum string) (bool, error) {
+	minV, err := parseMinimumBound(minimum)
+	if err != nil {
+		return false, err
+	}
+
+	maxExclusive, err := parseMaximumExclusiveUpper(maximum)
+	if err != nil {
+		return false, err
+	}
+
+	v, err := semver.NewVersion(trimSemverVPrefix(version))
+	if err != nil {
+		return maximum == "", nil
+	}
+
+	if minV != nil && v.LessThan(minV) {
+		return false, nil
+	}
+
+	if maxExclusive != nil && !v.LessThan(maxExclusive) {
+		return false, nil
+	}
+
+	return true, nil
+}
+
+func trimSemverVPrefix(s string) string {
+	return strings.TrimPrefix(strings.TrimSpace(s), "v")
+}
+
+func parseMinimumBound(minimum string) (*semver.Version, error) {
+	if minimum == "" {
+		return nil, nil
+	}
+
+	s := trimSemverVPrefix(minimum)
+	coerced, err := coerceSemverCore(s)
+	if err != nil {
+		return nil, fmt.Errorf("invalid minimum provided: '%s'", minimum)
+	}
+
+	parsed, err := semver.NewVersion(coerced)
+	if err != nil {
+		return nil, fmt.Errorf("invalid minimum provided: '%s'", minimum)
+	}
+
+	return parsed, nil
+}
+
+// parseMaximumExclusiveUpper returns the first version not allowed: valid versions satisfy v < upperExclusive.
+func parseMaximumExclusiveUpper(maximum string) (*semver.Version, error) {
+	if maximum == "" {
+		return nil, nil
+	}
+
+	s := trimSemverVPrefix(maximum)
+	core, tail := splitSemverCore(s)
+	parts := strings.Split(core, ".")
+	for _, p := range parts {
+		if _, err := strconv.ParseUint(p, 10, 64); err != nil {
+			return nil, fmt.Errorf("invalid maximum provided: '%s'", maximum)
+		}
+	}
+
+	switch len(parts) {
+	case 0, 1:
+		return nil, fmt.Errorf("invalid maximum provided: '%s'", maximum)
+	case 2:
+		if tail != "" {
+			return nil, fmt.Errorf("invalid maximum provided: '%s'", maximum)
+		}
+
+		maj, _ := strconv.ParseUint(parts[0], 10, 64)
+		min, _ := strconv.ParseUint(parts[1], 10, 64)
+
+		upper, err := semver.NewVersion(fmt.Sprintf("%d.%d.0-0", maj, min+1))
+		if err != nil {
+			return nil, fmt.Errorf("invalid maximum provided: '%s'", maximum)
+		}
+
+		return upper, nil
+	default:
+		full := core
+		if tail != "" {
+			full = core + tail
+		}
+
+		exclusive, err := semver.NewVersion(full)
+		if err != nil {
+			return nil, fmt.Errorf("invalid maximum provided: '%s'", maximum)
+		}
+
+		return exclusive, nil
+	}
+}
+
+func splitSemverCore(s string) (core, tail string) {
+	for i, r := range s {
+		if r == '-' || r == '+' {
+			return s[:i], s[i:]
+		}
+	}
+
+	return s, ""
+}
+
+func coerceSemverCore(s string) (string, error) {
+	core, tail := splitSemverCore(s)
+	parts := strings.Split(core, ".")
+	if len(parts) < 2 {
+		return "", fmt.Errorf("need at least major.minor")
+	}
+
+	for _, p := range parts {
+		if _, err := strconv.ParseUint(p, 10, 64); err != nil {
+			return "", err
+		}
+	}
+
+	for len(parts) < 3 {
+		parts = append(parts, "0")
+	}
+
+	out := strings.Join(parts, ".")
+	if tail != "" {
+		out += tail
+	}
+
+	return out, nil
 }
